Create folder /shared in /Checkout, and move the components in /components/checkout to this new folder. Delete the /components/checkout folder since that's now empty and won't be used
