+++ src/client/pages/Checkout/CheckoutDetails/CheckoutDetails.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import { useParams } from 'react-router-dom'
 import { useTextKeys } from 'utils/textKeys'
 import { useCurrentLocale } from 'l10n/useCurrentLocale'
-import { CheckoutPageWrapper } from 'components/checkout/CheckoutPageWrapper'
-import { Footer } from 'components/checkout/Footer'
+import { CheckoutPageWrapper } from '../shared/CheckoutPageWrapper'
+import { Footer } from '../shared/Footer'
 
 export const CheckoutDetails = () => {
   const textKeys = useTextKeys()
